remove feign client old code
